pyproject.toml
[tool.poetry]
name = "back-v3"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "back_v3"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.2"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.2.2"
python-dotenv = "^1.0.0"
google-auth = "^2.23.1"
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
